+++ CHANGELOG.md
@@ -5,3 +5,15 @@
 ## 1.0.5
 
 * Bug FIX
+
+## 1.0.6
+
+* Add optional field labelBackgroundColor param
+* Add optional field menuController param 
+* Correction itemBuilder tab to item
+
+## 1.0.7
+* Update screen when open menu is opened
+
+## 1.0.8
+* Use correct label background when widget is contained in a dialog